订单更新物流单，评价，显摆状态

+++ src/main/java/com/takefree/service/impl/OrderRateServiceImpl.java
@@ -1,11 +1,15 @@
 package com.takefree.service.impl;
 
+import com.takefree.enums.OrderRateEnum;
 import com.takefree.pojo.mapper.OrderRateMapper;
+import com.takefree.pojo.mapper.TakeOrderMapper;
 import com.takefree.pojo.model.OrderRate;
+import com.takefree.pojo.model.TakeOrder;
 import com.takefree.pojo.query.OrderRateQuery;
 import com.takefree.service.OrderRateService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -17,15 +21,37 @@ public class OrderRateServiceImpl implements OrderRateService {
     @Autowired
     private OrderRateMapper orderRateMapper;
 
+    @Autowired
+    private TakeOrderMapper takeOrderMapper;
+
     @Override
+    @Transactional
     public OrderRate create(OrderRate orderRate) {
-        orderRateMapper.insertSelective(orderRate);
+        int row=orderRateMapper.insertSelective(orderRate);
+        if(row>0){
+            TakeOrder takeOrder=new TakeOrder();
+            takeOrder.setRateStatus(OrderRateEnum.YES.getCode());
+            takeOrder.setId(orderRate.getOrderId());
+            takeOrderMapper.updateByPrimaryKeySelective(takeOrder);
+        }
         return orderRate;
     }
 
     @Override
+    @Transactional
     public int deleteById(Long id) {
-        return orderRateMapper.deleteByPrimaryKey(id);
+        OrderRate orderRate=orderRateMapper.selectByPrimaryKey(id);
+        if(orderRate==null){
+            return 0;
+        }
+        int row=orderRateMapper.deleteByPrimaryKey(id);
+        if(row>0){
+            TakeOrder takeOrder=new TakeOrder();
+            takeOrder.setRateStatus(OrderRateEnum.NO.getCode());
+            takeOrder.setId(orderRate.getOrderId());
+            takeOrderMapper.updateByPrimaryKeySelective(takeOrder);
+        }
+        return row;
     }
 
     @Override
